requirements.txt
Django==5.0.2
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4 